fix(gcp/compute.ts): set boot disk protect to true to prevent accidental deletion
Previously, the `protect` flag on the boot disk was set to `false`, which meant that the boot disk could be accidentally deleted. This change sets the `protect` flag to `true`, which will prevent the boot disk from being deleted unless the `protect` flag is explicitly set to `false`. This change will help to prevent accidental data loss.