feat(nix): add Kubernetes tools option and include packages
This commit introduces a new option `enableKubernetes` in `overall-options.nix` to control the installation of Kubernetes-related tools.
It also adds a comprehensive list of Kubernetes tools, including Minikube, Kind, K3d, Helm, Kubectl, K9s, and more, to the `optional-packages-without-config.nix` file, conditioned on the new option. This allows users to easily enable and install a suite of essential Kubernetes utilities with a simple configuration change.