ci(flake.nix): add nix-casks flake input to manage macOS applications
This change adds `nix-casks` as a new input to the `flake.nix` file. This allows us to manage macOS applications using Nix, similar to how we manage other system packages.

The primary reason for this change is to standardize the management of applications on macOS and leverage the benefits of Nix for reproducibility and declarative configuration.

By adding `nix-casks`, we enable the installation of applications like `key-codes` and `vlc` through Nix, which will be reflected in the `configuration.nix` file. This consolidates application management within the Nix ecosystem.